+++ Classes/Core/IAHGear.m
@@ -63,7 +63,7 @@
 
  Note: Call success even if you are not doing anything here.
  */
-- (void)fetchAllUpdateForUser:(IAHUser *)user fromTime:(long)time success:(void (^)(NSMutableArray* updateArray))success failure:(void (^)(NSError* e))failure
+- (void)fetchAllUpdateForUser:(IAHUser *)user fromTime:(long long)time success:(void (^)(NSMutableArray* updateArray))success failure:(void (^)(NSError* e))failure
 {
     success(nil);
 }
@@ -77,7 +77,7 @@
 
  Note: Call success even if you are not doing anything here.
  */
-- (void)addReply:(IAHTicketReply *)reply byUser:(IAHUser *)user getUpdatesFromTime:(long) time success:(void (^)(NSMutableArray* updateArray))success failure:(void (^)(NSError* e))failure
+- (void)addReply:(IAHTicketReply *)reply byUser:(IAHUser *)user getUpdatesFromTime:(long long) time success:(void (^)(NSMutableArray* updateArray))success failure:(void (^)(NSError* e))failure
 {
     success(nil);
 }
